ck: tcp_rt module: change the _tcp_rt_stats item type to u64
to #32682818

_tcp_rt_stats is used to save the values returned from tcp_rt_stats.
These values were originally 64-bit, and now stored in u32, some larger
variables will overflow, so they are modified to 64-bit.

Acked-by: Dust Li <dust.li@linux.alibaba.com>

+++ net/ipv4/tcp_rt/tcp_rt.h
@@ -95,17 +95,17 @@ struct tcp_rt_stats {
 };
 
 struct _tcp_rt_stats {
-	u32 rt;
-	u32 number;
-	u32 drop;
-	u32 bytes;
-	u32 server_time;
-	u32 fail;
-	u32 packets;
-	u32 rtt;
-	u32 upload_time;
-	u32 upload_data;
-	u32 con_num;
+	u64 rt;
+	u64 number;
+	u64 drop;
+	u64 bytes;
+	u64 server_time;
+	u64 fail;
+	u64 packets;
+	u64 rtt;
+	u64 upload_time;
+	u64 upload_data;
+	u64 con_num;
 };
 
 int tcp_rt_output_init(int log_buf_num, int stats_buf_num,
